RSS Feature Addition
This update removed the interactive mode to move towards a config file only situation. It added support for some additional config properties such as site title, site url, timezone, and an rss boolean. There are primarily used in aid of generating an xml document for rss. RSS feed as is supports post creations, updates, and removals.